refactor(world): move widget types to src/world/widgets/
- Move WidgetState type from widgetValueStore.ts to src/world/widgets/widgetState.ts
- Move widget component-key definitions to src/world/widgets/widgetComponents.ts
- Derive component bucket shapes from IBaseWidget via Pick<> instead of hand-rolling
- Re-export WidgetState from widgetValueStore for back-compat
- Convert slot from arrow expression to function declaration

Amp-Thread-ID: https://ampcode.com/threads/T-019df514-1f1c-7764-aa92-76b65210a74d
